Add c-spiffe.h header file
